requirements.txt

pandas
scikit-learn
tldextract
whois
sentence-transformers
requests
gunicorn
librosa

flask>=3.0
flask-cors>=4.0
flask-limiter>=3.5
flask-jwt-extended>=4.6
pymongo>=4.6
werkzeug>=3.0
python-dotenv>=1.0
pytesseract pillow